a few renames and GetModuleScenarioTests
