+++ src/test/java/com/andrascsanyi/encyclopediagalactica/iam/infra/mappers/ModuleMapperShouldTests.java
@@ -5,6 +5,9 @@ import static org.assertj.core.api.Assertions.assertThat;
 import com.andrascsanyi.encyclopediagalactica.iam.contracts.ModuleInput;
 import com.andrascsanyi.encyclopediagalactica.iam.contracts.ModuleOutput;
 import com.andrascsanyi.encyclopediagalactica.iam.entities.Module;
+import com.andrascsanyi.encyclopediagalactica.iam.infra.mappers.datasets.ModuleInputToModuleMapperDataSet;
+import com.andrascsanyi.encyclopediagalactica.iam.infra.mappers.datasets.ModuleListToModuleOutputListDataSet;
+import com.andrascsanyi.encyclopediagalactica.iam.infra.mappers.datasets.ModuleToModuleOutputDataSet;
 import java.util.List;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ArgumentsSource;
